[Hotfix] Bugfix 2 (#42)
* [Improve][seatunnel-server] Update api style follow front-end & product-design

* [Improve][seatunnel-server] Update api style follow front-end & product-design

* [Improve][seatunnel-server] Bug fix

* [Improve][seatunnel-server] Update api style follow front-end & product-design (#12)

* [Improve][seatunnel-server] Update api style follow front-end & product-design

* [Improve][seatunnel-server] Update api style follow front-end & product-design

* [Improve][seatunnel-server] Bug fix

---------


* [Improve][seatunnel-server] Bugfix

* [Improve][seatunnel-server] Update api style follow front-end & product-design

* [Improve][seatunnel-server] Update api style follow front-end & product-design

* [Improve][seatunnel-server] Bug fix

* [Improve][seatunnel-server] Bugfix

* [Improve][seatunnel-server] merge conflicts

* [Improve][seatunnel-server] bug fix

* [Improve][seatunnel-server] bug fix

* [Improve][seatunnel-server] bug fix

---------

+++ seatunnel-server/seatunnel-app/src/main/java/org/apache/seatunnel/app/domain/request/user/BaseUserReq.java
@@ -28,10 +28,10 @@ public class BaseUserReq {
     private String username;
     @ApiModelProperty(value = "user password", required = true, dataType = "String")
     private String password;
-    @ApiModelProperty(value = "user status", required = true, dataType = "Byte")
+    @ApiModelProperty(value = "user status", required = true, dataType = "Integer")
     @NotNull
     private Byte status;
-    @ApiModelProperty(value = "user type", required = true, dataType = "Byte")
+    @ApiModelProperty(value = "user type", required = true, dataType = "Integer")
     @NotNull
     private Byte type;
 }
